Include command arguments in help
Add each argument's name and optionality in the usage line for the
command's help (both command line and markdown).

The args add metadata to the command as annotations. That metadata is
reversed to create a formated arguments string when rendering the help.

The markdown docs do not allow for custom rendering of the body.
Instead, we can hack the command usage to add the arg formated string
and append it to the usage. This only works because the commands have
already been added to their parent command. As it's a hack, it may stop
working any time we update cobra. The tests will fail if/when this hack
no longer works.