Disable local-path builds from Windows
We can't make builds from Window's sources reliable because of the file
permission missmatch with the linux system in the Docker container.
Windows execute permissions are granted based on the file extension
while Linux execute permissions are explicitly granted.

The best way for Windows users to build a container from source is to
commit the local source into a git repository. Git has special handling
to preserve the file permissions when checked out on a linux system.

As it stands, builds that don't require executing any of the source
files will work, while all other builds will fail. The inoperable builds
often fail in subtile and confusing ways. Worse, it's quite possible for
the build to pass and later fail at runtime when a new binary is invoked
for the first time.

Because we can't have confidence in local builds from Windows, we are
removing the capability.