Spike: invoke handlers via port-forwarding
No longer depend on cluster ingress. This requires that a Pod for the
handler be running and will not work with a handler scaled to zero. Or
at least the CLI would need to force the handler to scale up before
trying to invoke.

The invoke command is still only intended for one-off requests and
should not be relied upon for production workloads as it is both slow,
resource intensive and requires access to the Kubernetes API server.