Consume Container CRD
Container is a build CRD similar to Function and Application, however,
it does not have a build component. It's sole responsibility is to
resolve the provided image to the latest available image with the
digest.

The Container CRD can be consumed by Handlers or any other resource that
understand the .status.latestImage interface.