Selectively enable runtimes at compile time
Some runtimes are more mature than others. We should have a way to pick
which runtimes are enabled by default. There is a new ldflag named
`github.com/projectriff/cli/pkg/cli.cli_runtimes` which accepts a comma
separated list of runtimes names.

Runtimes that are not included will not be registered on the root
command.

The core and knative runtimes are currently enabled. The streaming
runtime is disabled.

Fixes #52