Selectively enable runtimes at compile time (#54)
Some runtimes are more mature than others. We should have a way to pick
which runtimes are enabled by default. There is a new ldflag named
`github.com/projectriff/cli/pkg/cli.cli_runtimes` which accepts a comma
separated list of runtimes names.

Runtimes that are not included will still be registered on the root
command, but hidden from doc and completion scripts.

The core and knative runtimes are currently enabled. The streaming
runtime is disabled.

Fixes #52