Log when waiting for a resource to become ready
When creating a resource the `--tail` flag will block until the resource
becomes ready (or not). Many resources have logs that will be displayed,
but not all. Even for resources will logs, it may take a moment for logs
to start being produced.

Now we log when waiting for a resource to become ready.