Add support for imagebindings.bindings.projectriff.io (#239)
This is a more complex resource than we have managed so far because we
need to populate multiple object references in a user friendly way.
Currently, we require the use to specify the fully qualified resource
(not a Kind) and use the discovery client to lookup the kind. This extra
step both validates the user value and will allow us to use short names
and partial matches in the future, similar to kubectl.

The list command reverses the object reference back into the
input string.

There is a bit of an impedance mismatch between the knative.dev/pkg
based clients and the projectriff/system based clients, but they can
work together.