Allow specifying the docker container's network for local builds
Pack 0.11 changed the default container network from host to bridged
for the analyze and export containers. This breaks local builds that use
a registry DNS entry defined in /etc/hosts (like our acceptance test
sometimes do).

This adds a new flag, --docker-network, when creating a function or
application with a local path to set the docker network. The flag is
hidden and not unit tested.